+++ NEWS.md
@@ -7,9 +7,15 @@
 
 ## Bug fixes and improvements
 
-* Define the geographic zones in the documentation
+* Define the geographic zones for measuring pollution in the documentation
 * Correct the date ranges for the values needed to declare pollution emergencies in the README
-* rename `get_latest_data` to `get_latest_imeca` and `get_zone_data` to `get_zone_imeca`
+* Messages about measuring stations no longer included in the index are now shown with `message()` instead of `warning()`
+* Messages about changes in the way the IMECA is computed are now shown with `message()` instead of `warning()`
+
+## Deprecated and Defunct
+* `get_latest_data` is deprecated. You should instead use `get_latest_imeca`. 
+* `get_zone_data` is deprecated. You should instead use `get_zone_imeca`.
+* `showWarnings` argument to `get_zone_imeca` was deprecated for `show_messages`
 
 
 # aire.zmvm 0.5.0